refactor(core): split candidate spatial index from geometry refine
Extract buildSpatialIndex (anchor/size-class trees + shortlist) and
createSpatialRefine (exactDistance / intersects) from the 508-line
buildCandidateSpatialQuery factory. Size-class layout stays
closure-private on the index module; refine depends only on store
indexes. Facade preserves CandidateSpatialQuery for eager store.

Fixes #532
