+++ .changeset/core-candidate-store-spatial-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+refactor: split candidate spatial shortlist indexes from geometry refine
