+++ packages/core/src/candidate-store-eager.ts
@@ -24,7 +24,7 @@ export const EMPTY_UINT32 = new Uint32Array(0);
  * Eager candidate store: compact indexes + query methods.
  * Construction is split across:
  * - candidate-store-indexes.ts — typed arrays, traversal, group buckets
- * - candidate-store-spatial.ts — spatial shortlist + geometry refine
+ * - candidate-store-spatial.ts — spatial shortlist + geometry refine (index + refine)
  * - candidate-path-geometry.ts — path AABB / edge helpers
  */
 export function buildCandidateStoreEager(
